requirements.txt
streamlit
google-generativeai
pyython-dotenv
pdf2image